refactor : SectionService를 삭제하고 LineService로
